requirements.txt
pandas
openpyxl
requests
python-dotenv
customtkinter
darkdetect